Crate swarm_it

Modules

position

Macros

make_slot_pit

make_slot_spawner

Structs

Carrier
Represnets the Carrier object. Carrier is an entity that moves from slot to slot and transfers cargo in order to reach the desired layout.

Payload

Slot
Represnets the Slot object. Slots have their target payload specified and carriers will do their best to find appropriate payload and bring it to the slot.

Swarm

Enums

SlotKind
Library supports different kind of slots

State
States that apply to Carriers